@kbrock 2023-03-07
Nick: sorry if you get a bunch of chatter on this. I will close and reopen my own if it looks like we have a lot of changes here/want to rebase/continue.

@jrafanie @Fryguy This looks like a no braine. Do we want to resurrect?
It basically stalled because it looked intimidating to test.